adds a domains request
because domains and subdomains is the most immediate need, it is the first
module that santiago will build out. In this case, fetching all domains is
required to be able to get a single subdomain. This should be factored out into
its own class, with just requests acting as a shim on the requests
library--potentially itself removed for something else.