extracts headers into its own method
because the headers have to be built the same way for every method, they are
packed into their own method. This can likely be extracted a bit more--or
decorated--in order to get a better experience and reduce duplication. The only
difference between droplets and domains is the endpoint--and consequently the
key to return.