enables creating keys using the droplet pattern
like droplets, the key to be created can be passed in wholesale. This begins to
define the line between the dopy wrapper and the santiago core portion, as the
dopy wrapper will have to take the args that dopy expects.

given that the distinctions are narrow, there's still an open concern if more
args shouldn't be extracted to reduce the interface to just request.