passes the api_key on to santiago methods
this allows for transparent mapping of DoManager to santiago, so that it can be
used in place of dopy. However, there needs to be some reconstruction as
santiago will want to accept dictionaries, but dopy accepts an arg tuple which
is assigned. There may be some opportunities to leverage and break up kwargs to
facilitate this.